requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
drf-spectacular==0.20.2
drf-writable-nested==0.6.3
inflection==0.5.1
jsonschema==4.1.2
PyJWT==2.3.0
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
